Some experimenting with ideas
Played with the packaging here to get a better feel for it.
Still shaky on how all this shit fits together.

Also an alternate implementation that uses records instead
of subclasses, b/c the subclasses seem to fuck with the
type system a lot (I think it monomorphs vars, so they
take the first type they can identify, rather than the
base type that will work for all use cases. This means that
the order of expressions can change the types of variables
so sometimes I have to help it by declaring the type, or
by casting, or by just throwing my hands up and declaring
it Dynamic. Would be nice to not have to do that,
maybe the record style here will do the trick, IDK,
but thought it would also be nice to have these experiments
available

Run the two examples with:

$ haxe compile.hxml
$ haxe -x AlternateRuby.hx