Create an InternalMap class which is a haxe.ds.StringMap
IDK why there's all these different data structures,
but kinda like the idea of a layer of indirection between
my code and the ultimate thing. e.g. it might be useful
to actually make it a table of our symbols instead of
haxe Strings. Then again, it might not, IDK. At the very
least, the code is more aesthetically pleasing.