Rename ruby.WorldWorker -> ruby.World
It manages the common interactions to ruby.ds.World

Not sure if I like this or not.
Had a really hard time figuring out what I wanted to name it.
I expect it to grow lots of behaviour that will need to be extracted out
into their own objects that manage subdomains of functionality.

Also, put state back onto interpreter.
Also not sure if I like this or not, it was mostly done b/c it was
easier while I was moving things around.