Interpreter parses JSON null as "None" node
Using `nil` was causing it to go interpret that code and find intermediate values.
This was causing fkn confusing errors, e.g. the code `Object` would evaluate to
Constant(Nil, "Object"), it would then evaluate the `Nil` to get world.rubyNil,
which is an object, so intstead of returning world.objectClass, it returned
first world.rubyNil from the ns evaluation, and then world.objectClass from the
constant evaluation.