Remove AST
Translate incoming JSON directly into ExecutionStates instead of keeping
the intermediate ASTs. This is because as I wanted to create executions
from within other code, I wound up needing to generate ASTs instead of
Executions. Not totally sure this is the right thing to do, but it makes
this easier, and should retain an equivalent amount of information as
they're all in a "starting state", and they are immutable.

Also, saves from the tedious translation between ASTs and
ExecutionStates, which was basically 1-to-1, so one more place to have
to get the translation correct, but had no obvious benefit, and was
leading to "match anything" cases b/c more of AST was implemented than
ExecutionState, but now it's much more explicit, so can take advantage
of the pattern matching to see all unmatched expressions.
(currently `SetIvar(_) | GetIvar(_) | Float(_) | Integer(_): throw...`,
and a few states of other matches)

Methods store their signature in an array of ArgTypes instead of ASTs.
This allows them to be abstracted into something easy to work with,
instead of something tied to the code that defined them.

ExecutionState.OpenClass got built out a bit more.