Split String up to two expressions to see how easy/hard it is... swear this is a bug in Haxe
When I name the Enum value "Return", it can't find it and just returns null instead of blowing up like a reasonable implementation,
so, of course, it takes me forever to figure out that this is why its failing. When I rename the enum value from Return to ThenReturn,
it is able to find it in the hash again.