Refactor defn parser: consolidating paths through it
Its strange b/c the toplevel parsing is not nested within a node.
This one adds a toplevel root node so that it doesn't have to have so many special cases.
Also moves the parsing of the machine into its own object instead of its
parent. This way, it can modify its state as it parses instructions.